refactor(pdf): remove pdfminer extraction modules
Delete pdfminer processing, utility, compatibility, and patch modules while keeping layout cleanup and metadata paths on core-pdf records.
