+++ pyproject.toml
@@ -43,7 +43,6 @@ dependencies = [
     "tqdm>=4.67.3, <5.0.0",
     "typing-extensions>=4.15.0, <5.0.0",
     "unstructured-client>=0.25.9, <1.0.0",
-    "wrapt>=2.1.1, <3.0.0",
     "filelock>=3.12.0,<4.0.0",
 ]
 
@@ -66,7 +65,6 @@ image = [
     "core-pdf",
     "google-cloud-vision>=3.12.1, <4.0.0",
     "pdf2image>=1.17.0, <2.0.0",
-    "pdfminer.six>=20251230, <20270000",
     "pi-heif>=1.2.0, <2.0.0",
     "pikepdf>=10.3.0, <11.0.0",
     "pypdf>=6.6.2, <7.0.0",
