+++ unstructured/partition/utils/constants.py
@@ -4,7 +4,6 @@ from enum import Enum
 
 class Source(Enum):
     CORE_PDF = "core_pdf"
-    PDFMINER = "pdfminer"
     OCR_TESSERACT = "ocr_tesseract"
     OCR_PADDLE = "ocr_paddle"
     OCR_GOOGLEVISION = "ocr_googlevision"
